+++ packages/skirmish/vitest.config.js
@@ -1,0 +1,17 @@
+import { defineConfig } from "vitest/config";
+import { playwright } from "@vitest/browser-playwright";
+
+export default defineConfig({
+  test: {
+    include: ["tests/*.spec.res.mjs"],
+    browser: {
+      provider: playwright(),
+      enabled: true,
+      // at least one instance is required
+      instances: [{ browser: "chromium" }],
+    },
+    coverage: {
+      provider: "istanbul",
+    }
+  },
+});